backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Resume parsing
pdfminer.six==20221105
docx2txt==0.8
Pillow==10.1.0
pytesseract==0.3.10

# NLP and ML
sentence-transformers==3.0.1
transformers==4.45.0
torch==2.6.0
scikit-learn==1.5.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
